funciona perro

+++ UniversidadEquipo1/src/entidades/Inscripcion.java
@@ -55,7 +55,7 @@ public class Inscripcion {
         this.materia = materia;
     }
 
-    public double getNota() {
+    public int getNota() {
         return nota;
     }
